refactor: 개별 운행 정보 카드에서 스타일링 분리
